set up failing acceptance spec for update commands
There are a couple of things that end up happening here;

Most notably, there's no particular way to actually filter the remove
commands, so the first failure case ends up removing _all_ tasks.
This can arguably be fixed from within the execute method of the remove
task, but most command bases don't provide an easy way to _internally_
modify the relation that gets used by the execute method.

There's a related filtering problem where updates are concerned, which
should be exercised and demonstrated by this spec as well, but the
specific failure is masked by the unfiltered delete.